soc: overlay: Move ethernet overlay from soc to base
soc overlay is used for adding new node which are not
available in the base DTSI. Base overlay is used for
modifying the node properties of the node available in
the base node.

As ethernet node is already available in base dtsi,
move the overlay from soc-overlay to base-overlay file.

Bug 4057304

Change-Id: Ie7a2a49de15ec69d69b6839973123f467443c314
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/soc/t23x-generic-dts/+/2884656
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>