UPSTREAM: arm64: tegra: Add Jetson Orin NX support
This adds a device tree for the Jetson Orin NX module, which is Jetson
AGX Orin's little sibling with 6 or 8 ARM Cortex-A78AE cores, an Ampere
GPU (1024 GPU and 32 tensor cores) and a number of accelerators for
machine learning, image processing and more.

The Jetson Orin NX comes with either 8 or 16 GiB of 128-bit LPDDR5 and
supports NVME for mass storage.

Ref CL: https://git.kernel.org/pub/scm/linux/kernel/git/tegra/linux.git/commit/?h=for-6.4/arm64/dt&id=13b0aca303e97500e6aa6fa7965c620394c67775

Bug 3998955

Change-Id: I10aeb14293657e491e2d7627d7f2a3639b844be4
Signed-off-by: Thierry Reding <treding@nvidia.com>
Signed-off-by: Gautham Srinivasan <gauthams@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t23x-public-dts/+/2894102
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>