t23x: overlay: Disable rail-gating for safety IGX
This change disables GPU rail-gating as it's not a mandatory feature for
safety IGX. The ideal approach for disabling GPU rail-gating should be
deleting the `power-domains` property in GPU node. But the
/delete-property/ is not a valid syntax in the device tree overlay, the
`nvidia,tegra-joint_xpu_rail` is specified to achieve the same as an
alternative.

Bug 4084478

Signed-off-by: Yi-Wei Wang <yiweiw@nvidia.com>
Change-Id: Ia186cc0bc18f5d1a216a511af3e5cc4588c07b21
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t23x-public-dts/+/2946710
Reviewed-by: Sagar Kamble <skamble@nvidia.com>
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Rajkumar Kasirajan <rkasirajan@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>