overlay: put runtime fragments in separate overlay
This is an incremental step toward having a "with-oot" base dtb
that contains both the upstream dtb as well as the nvidia-oot
data in a single statically built dtb.

Note that currently the base dtb is stored and managed in the
rootfs via extlinux.conf file.  The overlays however live inside
the UEFI partition.  The ultimate goal is to have consistency
in how the dtb files are managed.

After we combine the data from nvidia-oot overlay dtb into the
future "with-oot" base dtb then we can move the remaining overlays
to the rootfs and manage all dtbs/overlays there.

This patch takes the first critical step toward this goal by
separating the static overlay data from the dynamic overlay data
that gets applied conditionally at run-time.

Bug 4290389

Change-Id: I403ac84b0737368b8bc96952552729ab7e46802b
Signed-off-by: Brad Griffis <bgriffis@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t23x-public-dts/+/2991524
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>