overlay: p3768: enable i2c5
UEFI expects that i2c5 is enabled and will throw an assert otherwise.
In k5.10 dtb this was enabled there as well, so enabling here for
consistency and to avoid issues with UEFI.

The i2c5 controller is necessary for HDMI use case but it is ok to
have it enabled for DP use case.

Bug 4327032

Change-Id: Ib9af7c95198b7650df2537d0859ed8a5b8af1aa0
Signed-off-by: Brad Griffis <bgriffis@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t23x-public-dts/+/2995550
Reviewed-by: Gautham Srinivasan <gauthams@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>